set hierarchy
