+++ classes/Engine/dieselEngine.java
@@ -1,0 +1,10 @@
+package classes.Engine;
+
+public class dieselEngine extends Engine{
+
+    private String fuelType;
+
+    public dieselEngine(String engineName, int engineMaxSpeed) {
+        super(engineName, engineMaxSpeed);
+    }
+}
