+++ classes/Engine/gasEngine.java
@@ -1,0 +1,10 @@
+package classes.Engine;
+
+public class gasEngine extends Engine{
+
+    private double compressionRatio;
+
+    public gasEngine(String engineName, int engineMaxSpeed) {
+        super(engineName, engineMaxSpeed);
+    }
+}
