+++ classes/Engine/petrolEngine.java
@@ -1,0 +1,10 @@
+package classes.Engine;
+
+public class petrolEngine extends Engine{
+
+    private double gasMileage;
+
+    public petrolEngine(String engineName, int engineMaxSpeed) {
+        super(engineName, engineMaxSpeed);
+    }
+}
