--- classes/PassengerCar.java
@@ -1,8 +1,0 @@
-package classes;
-
-public class PassengerCar extends Transport {
-
-    public PassengerCar(String name, Engine engine, Wheels wheels) {
-        super(name, engine, wheels);
-    }
-}
