--- classes/SportCar.java
@@ -1,8 +1,0 @@
-package classes;
-
-public class SportCar extends Transport{
-
-    public SportCar(String name, Engine engine, Wheels wheels) {
-        super(name, engine, wheels);
-    }
-}
