+++ classes/Transport/PassengerCar.java
@@ -1,0 +1,12 @@
+package classes.Transport;
+
+import classes.Engine.Engine;
+import classes.Transport.Transport;
+import classes.Wheels.Wheels;
+
+public class PassengerCar extends Transport {
+
+    public PassengerCar(String name, Engine engine, Wheels wheels) {
+        super(name, engine, wheels);
+    }
+}
