+++ classes/Transport/SportCar.java
@@ -1,0 +1,12 @@
+package classes.Transport;
+
+import classes.Engine.Engine;
+import classes.Transport.Transport;
+import classes.Wheels.Wheels;
+
+public class SportCar extends Transport {
+
+    public SportCar(String name, Engine engine, Wheels wheels) {
+        super(name, engine, wheels);
+    }
+}
