--- classes/Truck.java
@@ -1,13 +1,0 @@
-package classes;
-
-public class Truck extends Transport{
-    private int weight;
-
-    public Truck(String name, Engine engine, Wheels wheels) {
-        super(name, engine, wheels);
-    }
-
-    public void setWeight(int weight) {
-        this.weight = weight;
-    }
-}
