+++ classes/Wheels/Wheels.java
@@ -1,0 +1,15 @@
+package classes.Wheels;
+
+public class Wheels {
+    private String wheelsName;
+    private double wheelCoeffFriction;
+
+    public Wheels(String wheelsName, double wheelCoeffFriction) {
+        this.wheelsName = wheelsName;
+        this.wheelCoeffFriction = wheelCoeffFriction;
+    }
+
+    public double getWheelCoeffFriction() {
+        return wheelCoeffFriction;
+    }
+}
