+++ classes/Wheels/summerWheels.java
@@ -1,0 +1,12 @@
+package classes.Wheels;
+
+import classes.Wheels.Wheels;
+
+public class summerWheels extends Wheels {
+
+    private double coeffWeight;
+
+    public summerWheels(String wheelsName, double wheelCoeffFriction) {
+        super(wheelsName, wheelCoeffFriction);
+    }
+}
