+++ classes/Wheels/winterWheels.java
@@ -1,0 +1,10 @@
+package classes.Wheels;
+
+public class winterWheels extends Wheels{
+
+    private double spikeCoeff;
+
+    public winterWheels(String wheelsName, double wheelCoeffFriction) {
+        super(wheelsName, wheelCoeffFriction);
+    }
+}
